+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # tiledesk-dashboard
 
+### 2.3.22-rc10
+- Adds a check that saves the intent if it has no special characters
+- Replaces the name of actions with their icon
+- Adds the "Change department" action
+- Adds the "Replace bot" action
+- Bug Fixing
+
 ### 2.3.22-rc9
 - Add the ability to enable/disable the intent webhook
 - Fixes action bugs